requirements.txt
# Budget Assistant - RAG Chatbot
# Dependencies for AWS Bedrock Knowledge Base integration

flask>=3.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
